requirements.txt
# Pixel Refiner
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
